Fix more bugs around seafaring, especially related to persistence.

+++ test/__init__.py
@@ -15,6 +15,7 @@ import unittest
 class WidelandsTestCase(unittest.TestCase):
     do_use_random_directory = True
     path_to_widelands_binary = None
+    keep_output_around = False
 
     def setUp(self):
         if self.do_use_random_directory:
@@ -22,8 +23,10 @@ class WidelandsTestCase(unittest.TestCase):
         else:
             self.run_dir = os.path.join(tempfile.gettempdir(), "widelands_regression_test", self.__class__.__name__)
             if os.path.exists(self.run_dir):
-                shutil.rmtree(self.run_dir)
-            os.makedirs(self.run_dir)
+                if not self.keep_output_around:
+                    shutil.rmtree(self.run_dir)
+            else:
+                os.makedirs(self.run_dir)
         print "stdout: %s/stdout.txt" % self.run_dir
         self.widelands_returncode = 0
 
@@ -32,18 +35,19 @@ class WidelandsTestCase(unittest.TestCase):
         unittest.TestCase.run(self, result)
 
     def tearDown(self):
-        if self.currentResult.wasSuccessful():
+        if self.currentResult.wasSuccessful() and not self.keep_output_around:
             shutil.rmtree(self.run_dir)
 
     def run_widelands(self, **kwargs):
         """Runs widelands with the arguments given, catching stdout and stderr in files."""
-
         stdout_filename = os.path.join(self.run_dir, "stdout.txt")
+        if (os.path.exists(stdout_filename)):
+            os.unlink(stdout_filename)
 
         with open(stdout_filename, 'a') as stdout_file:
             args = [self.path_to_widelands_binary, '--verbose=true',
                     '--datadir=.', '--homedir=%s' % self.run_dir,
-                    '--disable_fx=true', '--disable_music=true', '--maxfps=25', ]
+                    '--disable_fx=true', '--disable_music=true' ]
             args += [ "--%s=%s" % (key, value) for key, value in kwargs.iteritems() ]
             stdout_file.write("---- TestRunner: Starting Widelands: %s\n\n" % args)
